explicit-comms: moved shuffle to its own file
